pyproject.toml
[project]
name = "p1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "groq>=0.37.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "lxml>=6.0.2",
    "mistralai>=1.12.4",
    "notebook>=7.5.5",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]
